+++ spring-terminator-api/src/test/java/com/dwarfeng/springterminator/api/integration/example/TelqosExample.java
@@ -1,5 +1,6 @@
 package com.dwarfeng.springterminator.api.integration.example;
 
+import com.dwarfeng.dutil.basic.io.CT;
 import com.dwarfeng.springterminator.stack.handler.TerminateHandler;
 import org.springframework.context.support.ClassPathXmlApplicationContext;
 
@@ -21,19 +22,19 @@ public class TelqosExample {
         TerminateHandler terminateHandler = ctx.getBean(TerminateHandler.class);
 
         // 显示欢迎信息。
-        System.out.println("开发者您好!");
-        System.out.println("这是一个示例, 用于演示 dwarfeng-spring-terminator 的功能");
-        System.out.println("您可以使用 telnet 客户端工具访问本机 ${telqos.port} 端口以体验本示例的功能");
-        System.out.println("您可以使用 telqos 的 shutdown 命令以关闭本示例...");
+        CT.trace("开发者您好!");
+        CT.trace("这是一个示例, 用于演示 dwarfeng-spring-terminator 的功能");
+        CT.trace("您可以使用 telnet 客户端工具访问本机 ${telqos.port} 端口以体验本示例的功能");
+        CT.trace("您可以使用 telqos 的 shutdown 命令以关闭本示例...");
 
         // 等待 terminateHandler 返回退出代码和重启标记。
         int exitCode = terminateHandler.getExitCode();
         boolean restartFlag = terminateHandler.getRestartFlag();
 
-        System.out.println("退出代码: " + exitCode);
-        System.out.println("重启标记: " + restartFlag);
-        System.out.println("本示例不演示重启功能, 将会直接使用 0 作为退出代码退出程序");
-        System.out.println("示例演示完毕, 感谢您测试与使用!");
+        CT.trace("退出代码: " + exitCode);
+        CT.trace("重启标记: " + restartFlag);
+        CT.trace("本示例不演示重启功能, 将会直接使用 0 作为退出代码退出程序");
+        CT.trace("示例演示完毕, 感谢您测试与使用!");
 
         System.exit(exitCode);
     }